First update for 25/4/24.
Updated DEVNOTES.md. Changed title to Wurdil - Dumpster Fire Edition. Updated App.tsx. We can now enter letters to our Wurdil guess via the keyboard. The delete handler is also working finally. Gotta make deep copies of nested arrays to overcome type errors. Was attempting to update two pieces of state in the same handler where the states rely on each other. Initially, was using if-else inside each setter and this was failing miserably. Changed it so that the logic is outside the setters but inside the handler and all working well.