Updated app.tsx.
A lot of work done on enterKeyHandler. We can hit enter and find out if our answer is correct, not a valid answer (not in the answer list), or a valid answer but incorrect. Other parts of state are updated accordingly. Added react-toastify for notifications. Updated how we store our strings/letters - everything is lower case but converted to upper case to be displayed only.