Add keyboardState and updateKeyboard func
keyboardState is an array of objects, each object has key and color props. The color is the color that will be displayed on the onscreen keyboard. updateKeyboard is run when the user submits an answer - the purpose is to update the keyboardState which in turn is used to render the background color for each key on the onscreen keyboard. This provides feedback to the user so they can play Wurdil.