feat: guess tiles indicate if each letter is present in the answer, in the correct position, or absent from the answer
Updated App.tsx - answer is passed to the Guess component. Updated Guess component. Each letter in answer is checked against the corresponding letter in the guess. Based on the results of the comparison, a position variable is set to 'correct,' 'absent,' or 'incorrect'. This position variable is passed to the Tile component. The Tile component uses the position prop to determine what the background color of each tile should be. This is done by dynamically adding a class to the appropriate html element.