add: add array of valid words to data.ts
Renamed words array to answers. The answer is randomly selected from
answers when the game starts. Words that appear in the 'valid' array
are now used to check if the players guess is valid. The user can now
guess words like 'adieu' or 'twill' or plurals of 4 letter words, but
the words in the 'valid' array won't be used to select answers.