feat: user can close game over + view board
Add NewGameButton.tsx which will allow reuse of the new game button in
both the App.tsx and in the GameOver.tsx component. The user can click
the 'close' button to close the GameOver component and see the game
area. Add isGameOverVisible state. setIsGameOverVisible used by
GameOver component in the close button to close GameOver.tsx.
isGameOverVisible used in App.tsx to conditionally render either the
GameOver screen or the NewGameButton component.