refactor: add reusable button component
Button component is used for new game and close button in Gameover.tsx
and for the new game button in App.tsx. It wasn't designed with reuse in
mind but more so because I wanted the user to be able to close the
GameOver component. It's not pretty but it works.