test: add code to click onscreen keyboard button
Spent time trying to stub and spy on getRandomArbitrary which is used
to choose a random answer from the answers array when the app loads. My
intention was to force the app to 'randomly' choose the same word each
time and then I could test success and failure paths. Was not able to
achieve this. Further, the Cypress '.type()' command requires a specific
element type (input, textarea, etc). I can't simulate keypresses using
type. We can still enter values for our guesses using our onscreen
keyboard, and so I added a small amount of code for this.