fix: remove test flake by adding try-catch
Was getting an AbortError in my test due to the test runner opening the
Information window, causing audio to play, and then immediately closing
it, causing the audio to be paused. The play promise hadn't always
resolved by the time we tried to pause and so the test was flaky.
Added a try-catch around the code that attempts to pause the audio in
our app.