test: add to our wurdil.cy.ts file
Added steps to test entering values, submitting a partial answer,
backspace functionality, submitting an empty answer, submitting an
answer that is not a word, submitting an answer that is a valid word but
is not the correct answer, attempting to delete after submitting an
answer that has been accepted, losing the game by submitting 6 incorrect
answers. Added custom command, pressKey, to make interacting with
onscreen keyboard easier.