test: add more checks in our test
Update DEVNOTES.md. Add more data-testids for testing. Add test for
the 'guess the first word!' prompt that appears when the game starts.
Add checks that: each guess has 5 tiles, game over component doesn't
exist during play, it is displayed when the game is over, the toast
message is displayed with the correct message when we lose, we can start
a new game by clicking a button, the guess the word prompt is shown
again, and all guess tiles are empty when the game starts.