fix: handle play() request interruption
The DOMException AbortError returned today in testing. I could solve by
increasing the .wait() to 3 seconds. Changed useEffect so that our play
promise is stored in a variable. We can check if that variable is not
undefined in the cleanup function and catch the error or pause the audio
as approprite. Add test to ensure each alphabetic key is working.