fix: accessability and focus/keyboard navigation
Info icon to open Information.tsx is now wrapped in a button. Each Key
component is now a button rather than a div. In enterKeyHandler, the
enter key only works to submit an answer if the current target element
does NOT have an onclick function. In other words, if you have a button
focussed and you hit the physical enter key, the onclick function of the
focussed element will be fired and we will not attempt to submit an
answer.