fix: open and close information works as expected
Moved close button/icon for Information component higher in render
statement so the close button is the first element encountered when
using keyboard navigation. In useTrapFocus.ts, set close button as
focus as soon as the Information component is visible. Keyboard navigation
and enter key is now working as it does on NYT site.